requirements.txt
appdirs==1.4.0
bcrypt==2.0.0
beautifulsoup4==4.5.3
blinker==1.4
cffi==1.9.1
coverage==4.3.4
Django==1.9.7
django-configurations==2.0
django-cors-middleware==1.3.1
django-filter==1.0.1
django-nose==1.4.4
djangorestframework==3.1.3
django-rest-framework-signature==1.0.1.dev2
django-rest-framework-simplify==1.0.1.dev
Markdown==2.6.8
mongoengine==0.9.0
mysqlclient==1.3.9
packaging==16.8
psycopg2==2.7.3.2
pymongo==3.2.1
pycparser==2.17
pycrypto==2.6.1
pyparsing==2.1.10
pymssql==2.1.3
python-dateutil==2.6.0
pytz==2016.1
raven==6.1.0
requests==2.4.3
six==1.10.0